# src/main/java/com/tnt/aggregator/rest/api/DetailsApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.2.3).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.tnt.aggregator.rest.api;

import com.tnt.aggregator.rest.Exception.ApiException;
import com.tnt.aggregator.rest.model.Details;
import io.swagger.annotations.*;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

import javax.validation.constraints.NotNull;
import javax.validation.constraints.Pattern;
import java.util.List;

/**
 * Auto-generated class using Swagger Maven plugin
 */
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2020-07-07T16:29:38.570+02:00")

@Api(value = "default", description = "the default API")
@RestController
public interface DetailsApi {

    @ApiOperation(value = "The API accepts three different parameters to specify the values to be passed to the individual backing APIs.", notes = "The API accepts three different parameters to specify the values to be passed to the individual backing APIs.", response = Details.class, tags={ "get details", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = Details.class),
        @ApiResponse(code = 400, message = "Invalid input", response = Void.class),
        @ApiResponse(code = 404, message = "Details not found", response = Void.class),
        @ApiResponse(code = 503, message = "Service unavailable", response = Void.class) })
    
    @RequestMapping(value = "/aggregation",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Details> getPackageDetails( @NotNull@ApiParam(value = "county codes", required = true) @RequestParam(value = "pricing", required = true) List<String> pricing, @NotNull@ApiParam(value = "9 digit order numbers", required = true) @RequestParam(value = "track", required = true) List<String> track, @NotNull@ApiParam(value = "9 digit order numbers", required = true) @RequestParam(value = "shipments", required = true) @Pattern(regexp = "[123456789]") List<String> shipments) throws ApiException;


}